requirements.txt
Pillow==9.4.0
numpy==1.23.5
colorama==0.4.6
pyfiglet==0.8.post1
